Create a PlanToSpec provider for CAMP.
Move all CAMP-related functionality from core into brooklyn-camp. Core doesn't know the specifics of plan providers now.
