+++ core/src/main/java/brooklyn/plan/PlanNotRecognized.java
@@ -1,0 +1,40 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package brooklyn.plan;
+
+public class PlanNotRecognized extends RuntimeException {
+    private static final long serialVersionUID = 1L;
+
+    public PlanNotRecognized() {
+        super();
+    }
+
+    public PlanNotRecognized(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public PlanNotRecognized(String message) {
+        super(message);
+    }
+
+    public PlanNotRecognized(Throwable cause) {
+        super(cause);
+    }
+
+}
