+++ usage/camp/src/test/java/brooklyn/camp/lite/TestAppAssembly.java
